Convert method
Converted all my ArrayByteInputStream into FileStream because the ArrayByteInputStream was not serializable, which mean after joining into another DFS, it was not able to put because the IS needed to be serializable. I tweaked the FileStream a bit to accept an Byte array just to make it easier on us. I've tested it quite a bit and now it should be working.